Bug in AFP font metrics that performed integer arithmetic when it should have been double arithemetic.
Also added some generics to the surrounding classes


git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@1311638 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/afp/fonts/CharacterSetBuilder.java
@@ -285,12 +285,12 @@ public abstract class CharacterSetBuilder {
                 CharacterSetOrientation[] characterSetOrientations
                     = processFontOrientation(structuredFieldReader);
 
-                int metricNormalizationFactor;
+                double metricNormalizationFactor;
                 if (fontControl.isRelative()) {
                     metricNormalizationFactor = 1;
                 } else {
                     int dpi = fontControl.getDpi();
-                    metricNormalizationFactor = 1000 * 72000
+                    metricNormalizationFactor = 1000.0d * 72000.0d
                         / fontDescriptor.getNominalFontSizeInMillipoints() / dpi;
                 }
